consolidate consultar RZ tests
